Small changes regarding formatting - Fixed formatting issues - Added comment block in beginning of script - Fixed typos - Renamed some variables

+++ Smart Updates/Reboot-HVClusterSU.ps1
@@ -1,100 +1,106 @@
-# Get all VM in the Cluster
-$Servers = Get-ClusterResource | Where-Object{$_.ResourceType -like "Virtual Machine"}
-# Get all prefered HyperV Nodes in the Cluster
-$preferedHosts=Get-ClusterGroup | Get-ClusterOwnerNode
-$file = $MyInvocation.MyCommand.Path
-# Check if Alle Clusternodes ere running and exit Skript if not all are running
-$running=Get-Clusternode  | Where-Object State -NotContains "Up"
+<#
+    .SYNOPSIS
+    Reboot HyperV Cluster Node and move all VMs to other HyperV Nodes.
+    
+    .DESCRIPTION
+    This script will reboot the HyperV Cluster Node and move all VMs to other HyperV Nodes. After the reboot all VMs will be moved back to the original HyperV Node.
 
+    .EXAMPLE 
+    PS> .\Reboot-HVClusterSU.ps1
+
+    .NOTES
+    Author  : servereye
+    Version : 1.0
+#>
+
+#region Variables
+# Get all VMs in the Cluster
+$Servers = Get-ClusterResource | Where-Object { $_.ResourceType -like "Virtual Machine" }
+# Get all preferred HyperV Nodes in the Cluster
+$preferredHosts = Get-ClusterGroup | Get-ClusterOwnerNode
+$file = $MyInvocation.MyCommand.Path
+# Check if all Clusternodes are running and exit script if any are not
+$running = Get-Clusternode | Where-Object State -NotContains "Up"
 
 $EventSourceName = "ServerEye-Custom"
 $script:_SilentOverride = $true
 $script:_SilentEventlog = $true
 if (Test-Path "C:\ProgramData\ServerEye3") {
     $_LogFilePath = "C:\ProgramData\ServerEye3\logs\smartUpdates.Clusterrestart.temp"
-}else {
+} else {
     $_LogFilePath = "C:\ProgramData\ServerEye\logs\smartUpdates.Clusterrestart.temp"
 }
 
-$datastorage=Get-ChildItem "C:\ClusterStorage"
-$datastorage=$datastorage[0].FullName
-$skriptrunning="$datastorage\skriptrunning.txt"
+$datastorage = Get-ChildItem "C:\ClusterStorage"
+$datastorage = $datastorage[0].FullName
+$scriptrunning = "$datastorage\scriptrunning.txt"
+#endregion
 
+#region Functions
 function Write-Log {
     <#
-            .SYNOPSIS
-                A swift logging function.
-            
-            .DESCRIPTION
-                A simple way to produce logs in various formats.
-                Log-Types:
-                - Eventlog (Application --> ServerEyeDeployment)
-                - LogFile (Includes timestamp, EntryType, EventID and Message)
-                - Screen (Includes only the message)
-            
-            .PARAMETER Message
-                The message to log.
-            
-            .PARAMETER Silent
-                Whether anything should be written to host. Is controlled by the closest scoped $_SilentOverride variable, unless specified.
- 
-            .PARAMETER SilentEventlog
-                Whether anything should be written to the Eventlog. Is controlled by the closest scoped $_SilentEventlog variable, unless specified.
-            
-            .PARAMETER ForegroundColor
-                In what color messages should be written to the host.
-                Ignored if silent is set to true.
-            
-            .PARAMETER NoNewLine
-                Prevents Debug to host to move on to the next line.
-                Ignored if silent is set to true.
-            
-            .PARAMETER EventID
-                ID of the event as logged to both the eventlog as well as the logfile.
-                Defaults to 1000
-            
-            .PARAMETER EntryType
-                The type of event that is written.
-                By default an information event is written.
-            
-            .PARAMETER LogFilePath
-                The path to the file (including filename) that is written to.
-                Is controlled by the closest scoped $_LogFilePath variable, unless specified.
-                
-        #>
+        .SYNOPSIS
+            A swift logging function.
+        
+        .DESCRIPTION
+            A simple way to produce logs in various formats.
+            Log-Types:
+            - Eventlog (Application --> ServerEyeDeployment)
+            - LogFile (Includes timestamp, EntryType, EventID and Message)
+            - Screen (Includes only the message)
+        
+        .PARAMETER Message
+            The message to log.
+        
+        .PARAMETER Silent
+            Whether anything should be written to host. Is controlled by the closest scoped $_SilentOverride variable, unless specified.
+
+        .PARAMETER SilentEventlog
+            Whether anything should be written to the Eventlog. Is controlled by the closest scoped $_SilentEventlog variable, unless specified.
+        
+        .PARAMETER ForegroundColor
+            In what color messages should be written to the host.
+            Ignored if silent is set to true.
+        
+        .PARAMETER NoNewLine
+            Prevents Debug to host to move on to the next line.
+            Ignored if silent is set to true.
+        
+        .PARAMETER EventID
+            ID of the event as logged to both the eventlog as well as the logfile.
+            Defaults to 1000
+        
+        .PARAMETER EntryType
+            The type of event that is written.
+            By default an information event is written.
+        
+        .PARAMETER LogFilePath
+            The path to the file (including filename) that is written to.
+            Is controlled by the closest scoped $_LogFilePath variable, unless specified.
+    #>
     [CmdletBinding()]
     Param (
         [Parameter(Position = 0)]
-        [string]
-        $Message,
-            
-        [bool]
-        $Silent = $_SilentOverride,
- 
-        [bool]
-        $SilentEventlog = $_SilentEventlog,
+        [string] $Message,
         
-        [System.ConsoleColor]
-        $ForegroundColor,
-            
-        [switch]
-        $NoNewLine,
-            
+        [bool] $Silent = $_SilentOverride,
+
+        [bool] $SilentEventlog = $_SilentEventlog,
+        
+        [System.ConsoleColor] $ForegroundColor,
+        
+        [switch] $NoNewLine,
+        
         [Parameter(Position = 1)]
-        [int]
-        $EventID = 1000,
- 
+        [int] $EventID = 1000,
+
         [Parameter(Position = 1)]
-        [string]
-        $Source,
- 
-            
+        [string] $Source,
+        
         [Parameter(Position = 3)]
-        [System.Diagnostics.EventLogEntryType]
-        $EntryType = ([System.Diagnostics.EventLogEntryType]::Information),
-            
-        [string]
-        $LogFilePath = $_LogFilePath
+        [System.Diagnostics.EventLogEntryType] $EntryType = ([System.Diagnostics.EventLogEntryType]::Information),
+        
+        [string] $LogFilePath = $_LogFilePath
     )
   
     # Log to Eventlog
@@ -102,12 +108,11 @@ function Write-Log {
         try { Write-EventLog -Message $message -LogName 'Application' -Source $Source -Category 0 -EventId $EventID -EntryType $EntryType -ErrorAction Stop }
         catch { }
     }
- 
-        
+    
     # Log to File
     try { "$(Get-Date -Format "yyyy-MM-dd HH:mm:ss") $EntryType $EventID - $Message" | Out-File -FilePath $LogFilePath -Append -Encoding UTF8 -ErrorAction Stop }
     catch { }
-        
+    
     # Write to screen
     if (-not $Silent) {
         $splat = @{ }
@@ -116,113 +121,102 @@ function Write-Log {
         if ($PSBoundParameters.ContainsKey('NoNewLine')) { $splat['NoNewLine'] = $NoNewLine }
         Write-Output @splat
     }
- 
 }
 
-
 function move_vms {
-    $Servers = Get-ClusterResource | Where-Object{$_.ResourceType -like "Virtual Machine"}
-    $server = $Servers| Where-Object Ownernode -like $env:computername
-    $Servers = $Servers| Where-Object State -like "Online"
-    foreach ($server in $servers){
-        if ($server.OwnerNode -like $env:computername){
+    $Servers = Get-ClusterResource | Where-Object { $_.ResourceType -like "Virtual Machine" }
+    $server = $Servers | Where-Object Ownernode -like $env:computername
+    $Servers = $Servers | Where-Object State -like "Online"
+    foreach ($server in $servers) {
+        if ($server.OwnerNode -like $env:computername) {
             Write-Log -Source $EventSourceName -EventID 3002 -EntryType Info -Message "VM $server wird verschoben"
             Move-ClusterVirtualMachineRole -Name $server.OwnerGroup -MigrationType Live 
         }
     }
 }
+
 function move_vms_back {
-    $Servers = Get-ClusterResource | Where-Object{$_.ResourceType -like "Virtual Machine"}
-    $Servers = $Servers| Where-Object State -like "Online"
-    foreach ($server in $servers){
-        $preferedHost=$preferedHosts |Where-Object ClusterObject -Match $Server.OwnerGroup
-        $preferedHost=$preferedHost.OwnerNodes[0].Name #nimm immer den am meisten bevorzugten hyperv falls mehrere bevorzugt werden
-        if ($Server.OwnerNode.Name -notlike $preferedHost){
+    $Servers = Get-ClusterResource | Where-Object { $_.ResourceType -like "Virtual Machine" }
+    $Servers = $Servers | Where-Object State -like "Online"
+    foreach ($server in $servers) {
+        $preferredHost = $preferredHosts | Where-Object ClusterObject -Match $Server.OwnerGroup
+        $preferredHost = $preferredHost.OwnerNodes[0].Name # Always use preferred HyperV Node
+        if ($Server.OwnerNode.Name -notlike $preferredHost) {
             Write-Log -Source $EventSourceName -EventID 3002 -EntryType Info -Message "VM $server wird zurückgeschoben"
-            Move-ClusterVirtualMachineRole -Name $Server.OwnerGroup -Node $preferedHost -MigrationType Live
+            Move-ClusterVirtualMachineRole -Name $Server.OwnerGroup -Node $preferredHost -MigrationType Live
         }
     }
 }
-function set_preferred_owner{
-    foreach($preferedHost in $preferedHosts){
-        if(!$preferedHost.OwnerNodes){
-            $Server=Get-ClusterResource | Where-Object{$_.ResourceType -like "Virtual Machine"}|Where-Object OwnerGroup -Like $preferedHost.ClusterObject
+
+function set_preferred_owner {
+    foreach ($preferredHost in $preferredHosts) {
+        if (!$preferredHost.OwnerNodes) {
+            $Server = Get-ClusterResource | Where-Object { $_.ResourceType -like "Virtual Machine" } | Where-Object OwnerGroup -Like $preferredHost.ClusterObject
             Set-ClusterOwnerNode -Group $server.OwnerGroup -Owners $Server.OwnerNode.Name
         }
-
-        
     }
-
 }
+
 function Reboot_HV {
-    #sort by priority to shutdown
- 
-    #setting up switch case ($false=shutdown,$true=start)
+    # Setting up switch case ($false=shutdown, $true=start)
     $b = Test-Path 'HKLM:\SOFTWARE\ShutdownScript'
     
- 
     switch ($b) {
         $false {
             try {
-                If ($running){
+                if ($running) {
                     exit "not all HyperV nodes are running"
                 }
 
-                if (!(Test-Path $skriptrunning)) {
-                    New-Item -Name "skriptrunning.txt" -Path $datastorage
-                    Set-Content $skriptrunning -Value $env:computername
-                }else {
-                    $running= Get-Content $skriptrunning
+                if (!(Test-Path $scriptrunning)) {
+                    New-Item -Name "scriptrunning.txt" -Path $datastorage
+                    Set-Content $scriptrunning -Value $env:computername
+                } else {
+                    $running = Get-Content $scriptrunning
                     if (!$running) {
-                        Set-Content $skriptrunning -Value $env:computername    
-                    }else {
+                        Set-Content $scriptrunning -Value $env:computername    
+                    } else {
                         Write-Log -Source $EventSourceName -EventID 3002 -EntryType Error -Message "Skript läuft noch auf einem anderen HyperV"
                         exit
                     }
                 } 
-                #add and register scheduled task
+                # Add and register scheduled task
                 $para = "-ExecutionPolicy unrestricted -NonInteractive -WindowStyle Hidden -NoLogo -NoProfile -NoExit -File " + '"' + $file + '"'
                 $Action = New-ScheduledTaskAction -Execute 'powershell.exe' -Argument $para
                 $Option = New-ScheduledTaskSettingsSet -AllowStartIfOnBatteries -WakeToRun
                 $Trigger = New-JobTrigger -AtStartUp -RandomDelay (New-TimeSpan -Minutes 5)
                 Register-ScheduledTask -TaskName RebootHVResumeJob -Action $Action -Trigger $Trigger -Settings $Option -RunLevel Highest -User "System"
      
-                #create regkey
+                # Create regkey
                 New-Item -Path HKLM:\SOFTWARE\ -Name ShutdownScript -Force
                 
-                
-                
-                #loop for trying to move all vm to other hyperV
-                $vms =get-vm |Where-Object State -like "Running"
+                # Loop for trying to move all VMs to other HyperV Node
+                $vms = get-vm | Where-Object State -like "Running"
                 for ($i = 0; $i -lt 5; $i++) {
                     move_vms
-                    $vms =get-vm |Where-Object State -like "Running"
-                    $Servers = Get-ClusterResource | Where-Object{$_.ResourceType -like "Virtual Machine"}
-                    $servers = $Servers| Where-Object Ownernode -like $env:computername
-                    $Servers = $Servers| Where-Object State -like "Online"
+                    $vms = get-vm | Where-Object State -like "Running"
+                    $Servers = Get-ClusterResource | Where-Object { $_.ResourceType -like "Virtual Machine" }
+                    $servers = $Servers | Where-Object Ownernode -like $env:computername
+                    $Servers = $Servers | Where-Object State -like "Online"
                     if (!$servers) {
-                        foreach($vm in $vms.Name){
+                        foreach ($vm in $vms.Name) {
                             stop-vm -Name $vm
                             Start-Sleep -Seconds 30
                         }
-                        $vms =get-vm |Where-Object State -like "Running"
+                        $vms = get-vm | Where-Object State -like "Running"
                         if ($null -eq $vms) {
-                            $i=5
+                            $i = 5
                         }
-
-                        
                     }
-
                 } 
                 if ($null -ne $vms) {
-                    Write-Log -Source $EventSourceName -EventID 3002 -EntryType Error -Message "5 mal erfolglos versucht alle VMs zu verschieben, abbruch"
+                    Write-Log -Source $EventSourceName -EventID 3002 -EntryType Error -Message "5 mal erfolglos versucht alle VMs zu verschieben, Abbruch!"
                     move_vms_back
                     Exit "VMS konnten nicht verschoben werden"
                 }
                 Suspend-ClusterNode -Name $env:computername
-                #wait 3sec then reboot
                 Start-Sleep -s 3
-                #Reboot Hyper-V
+                # Reboot Hyper-V
                 $Comment = "Hyper V Reboot for Smart Updates"
                 $reason = "P"
                 $major = 0
@@ -239,35 +233,35 @@ function Reboot_HV {
                 Write-Log -Source $EventSourceName -EventID 3002 -EntryType Info -Message "Starte Smartupdates"
                 Start-Process $patchrun -ArgumentList "force" -Wait
                 Start-Process @startProcessParams  
-                
             }
             catch {
                 move_vms_back
                 Write-Log -Source $EventSourceName -EventID 3002 -EntryType Error -Message "Something went wrong $_ "
             }
-            
         }
         $true {
-            resume-ClusterNode -Name $env:computername
+            Resume-ClusterNode -Name $env:computername
 
             Start-Sleep -Seconds 60
             move_vms_back
-            Set-Content $skriptrunning -Value $null
-            #remove regkey
+            Set-Content $scriptrunning -Value $null
+            # Remove regkey
             Remove-Item -Path HKLM:\SOFTWARE\ShutdownScript -Recurse
-            #remove job 
+            # Remove job 
             Unregister-ScheduledTask -TaskName RebootHVResumeJob -Confirm:$False
             Write-Log -Source $EventSourceName -EventID 3002 -EntryType Info -Message "Skript erfolgreich"
         }
     }
 }
+#endregion
 
-
+#region Main execution
 set_preferred_owner
-$preferedHosts=Get-ClusterGroup | Get-ClusterOwnerNode
+$preferredHosts = Get-ClusterGroup | Get-ClusterOwnerNode
 Reboot_HV
 
-Test-Path ("C:\ProgramData\ServerEye3\logs\smartUpdates.Clusterrestart.log"){
+if (Test-Path "C:\ProgramData\ServerEye3\logs\smartUpdates.Clusterrestart.log") {
     Remove-Item "C:\ProgramData\ServerEye3\logs\smartUpdates.Clusterrestart.log" -Force
 }
 Rename-Item -Path "C:\ProgramData\ServerEye3\logs\smartUpdates.Clusterrestart.temp" -NewName "C:\ProgramData\ServerEye3\logs\smartUpdates.Clusterrestart.log"
+#endregion
